refactor: rename schemas.py (of api/v1) to schemas_v1.py
